requirements.txt
anthropic>=0.18.0
openai>=1.12.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
pyyaml>=6.0
